create-tarball: program to make a tarball including all the submodule files
The problem with this style of repo (many submodules) as a build source
is that if those submodules delete the target commits, the repo is no
longer buildable, leaving us with deployed artifacts that we can't
recreate.

This program is part of the solution: after tagging the repo for a new
cyruslibs-vXX release, create a tarball containing every file at its
current position. Upload the tarball, and then we can build from that.