libical: update to track cyrusimap/cyrus-libical-v4-with-patches
The situation is now as follows.  In x/y/z, x is the repository owner, y
is the repository name, and z is the branch.

libical/libical/master is the current state of libical development.  It
includes changes that break cyrus-imapd, and which we (the Cyrus dev
team) will be talking about with libical.

cyrusimap/libical/cyrus-libical-v4 is a commit from
libical/libical/master's history from before those breaking changes.  It
includes no original work from outside the libical/libical/master
history.

cyrusimap/libical/cyrus-libical-v4-with-patches is a descendant of
cyrusimap/libical/cyrus-libical-v4, including work requires for
cyrus-imapd's continued work.

cyrusimap/libical/cyrus is the branch we sync into cyruslibs.  At
present, it has the same referent as cyrus-libical-v4-with-patches.

All new commits for libical for cyrus-imapd will be added to
cyrus-libical-v4-with-patches until such time as we resolve the issues
preventing us from following libical/libical/master.