build.sh: run libical tests after install
We updated the build script to ran tests before install, but
libical tests link against the installed library, not the one
that just got built. We need to install before testing. This
might only be applicable for libicalvcard.

This should get fixed in the CMake build config, but until then
we'll run tests after install.